split out data_source panel. Some validation dropped
