+++ dp_wizard/shiny/panels/dataset_panel/__init__.py
@@ -1,12 +1,6 @@
-from pathlib import Path
-from typing import Optional
-
-from dp_wizard_templates.code_template import Template
 from shiny import Inputs, Outputs, Session, reactive, render, ui
 
-from dp_wizard import opendp_version, package_root
 from dp_wizard.shiny.components.icons import (
-    data_source_icon,
     product_icon,
     unit_of_privacy_icon,
 )
@@ -19,14 +13,12 @@ from dp_wizard.shiny.components.outputs import (
     only_for_screenreader,
     tutorial_box,
 )
-from dp_wizard.types import AppState, Product
-from dp_wizard.utils.argparse_helpers import (
-    PRIVATE_TEXT,
-    PUBLIC_PRIVATE_TEXT,
-    PUBLIC_TEXT,
+from dp_wizard.shiny.panels.dataset_panel.data_source import (
+    data_source_server,
+    data_source_ui,
 )
+from dp_wizard.types import AppState, Product
 from dp_wizard.utils.code_generators import make_privacy_unit_block
-from dp_wizard.utils.csv_helper import get_csv_names_mismatch, read_csv_names
 
 dataset_panel_id = "dataset_panel"
 
@@ -78,11 +70,7 @@ def dataset_ui():
         ui.output_ui("dataset_release_warning_ui"),
         ui.output_ui("welcome_ui"),
         ui.layout_columns(
-            ui.card(
-                ui.card_header(data_source_icon, "Data Source"),
-                ui.output_ui("csv_or_columns_ui"),
-                ui.output_ui("row_count_bounds_ui"),
-            ),
+            data_source_ui(),
             [
                 ui.card(
                     ui.card_header(unit_of_privacy_icon, "Unit of Privacy"),
@@ -116,10 +104,10 @@ def dataset_server(
     is_tutorial_mode = state.is_tutorial_mode
 
     # Dataset choices:
-    initial_private_csv_path = state.initial_private_csv_path
-    private_csv_path = state.private_csv_path
-    initial_public_csv_path = state.initial_public_csv_path
-    public_csv_path = state.public_csv_path
+    # initial_private_csv_path = state.initial_private_csv_path
+    # private_csv_path = state.private_csv_path
+    # initial_public_csv_path = state.initial_public_csv_path
+    # public_csv_path = state.public_csv_path
     contributions = state.contributions
     contributions_entity = state.contributions_entity
     max_rows = state.max_rows
@@ -142,42 +130,6 @@ def dataset_server(
 
     # Release state:
     released = state.released
-
-    @reactive.effect
-    @reactive.event(input.public_csv_path)
-    def _on_public_csv_path_change():
-        path = input.public_csv_path()[0]["datapath"]
-        public_csv_path.set(path)
-        column_names.set(read_csv_names(Path(path)))
-
-    @reactive.effect
-    @reactive.event(input.private_csv_path)
-    def _on_private_csv_path_change():
-        path = input.private_csv_path()[0]["datapath"]
-        private_csv_path.set(path)
-        column_names.set(read_csv_names(Path(path)))
-
-    @reactive.effect
-    @reactive.event(input.column_names)
-    def _on_column_names_change():
-        column_names.set(
-            [
-                clean
-                for line in input.column_names().splitlines()
-                if (clean := line.strip())
-            ]
-        )
-
-    @reactive.calc
-    def csv_column_mismatch_calc() -> Optional[tuple[set, set]]:
-        public = public_csv_path()
-        private = private_csv_path()
-        if public and private:
-            just_public, just_private = get_csv_names_mismatch(
-                Path(public), Path(private)
-            )
-            if just_public or just_private:
-                return just_public, just_private
 
     @render.ui
     def dataset_release_warning_ui():
@@ -212,144 +164,6 @@ def dataset_server(
                 """,
             ),
         )
-
-    @render.ui
-    def csv_or_columns_ui():
-        if in_cloud:
-            content = [
-                ui.markdown(
-                    """
-                    Provide the names of columns you'll use in your analysis,
-                    one per line, with no extra punctuation.
-                    """
-                ),
-                tutorial_box(
-                    is_tutorial_mode(),
-                    """
-                    When [installed and run
-                    locally](https://pypi.org/project/dp_wizard/),
-                    DP Wizard allows you to specify a private and public CSV,
-                    but for the safety of your data, in the cloud
-                    DP Wizard only accepts column names.
-
-                    If you don't have other ideas, we can imagine
-                    a CSV of student quiz grades: Enter `student_id`,
-                    `quiz_id`, `grade`, and `class_year_str` below,
-                    each on a separate line.
-                    """,
-                    responsive=False,
-                ),
-                ui.input_text_area("column_names", "CSV Column Names", rows=5),
-            ]
-        else:
-            content = [
-                ui.markdown(
-                    f"""
-Choose **Private CSV** {PRIVATE_TEXT}
-
-Choose **Public CSV** {PUBLIC_TEXT}
-
-Choose both **Private CSV** and **Public CSV** {PUBLIC_PRIVATE_TEXT}
-                    """
-                ),
-                ui.output_ui("input_files_ui"),
-                ui.output_ui("csv_column_match_ui"),
-            ]
-
-        content += [
-            code_sample(
-                "Context",
-                Template(
-                    # NOTE: If stats vs. synth is moved to the top of the flow,
-                    # then we can show the appropriate template here.
-                    "stats_context",
-                    package_root / "utils/code_generators/no-tests",
-                )
-                .fill_values(CSV_PATH="sample.csv")
-                .fill_expressions(
-                    MARGINS_LIST="margins",
-                    EXTRA_COLUMNS="extra_columns",
-                    OPENDP_V_VERSION=f"v{opendp_version}",
-                    WEIGHTS="weights",
-                )
-                .fill_code_blocks(
-                    PRIVACY_UNIT_BLOCK="",
-                    PRIVACY_LOSS_BLOCK="",
-                    OPTIONAL_CSV_BLOCK=(
-                        "# More of these slots will be filled in\n"
-                        "# as you move through DP Wizard.\n"
-                    ),
-                )
-                .finish()
-                .strip(),
-            ),
-            ui.output_ui("python_tutorial_ui"),
-        ]
-        return content
-
-    @render.ui
-    def input_files_ui():
-        # We can't set the actual value of a file input,
-        # but the placeholder string is a good substitute.
-        #
-        # Make sure this doesn't depend on reactive values,
-        # for two reasons:
-        # - If there is a dependency, the inputs are redrawn,
-        #   and it looks like the file input is unset.
-        # - After file upload, the internal copy of the file
-        #   is renamed to something like "0.csv".
-        return [
-            tutorial_box(
-                is_tutorial_mode(),
-                (
-                    """
-                    For the tutorial, we've provided the grades
-                    on assignments for a school class in `sample.csv`.
-                    You don't need to upload an additional file.
-                    """
-                    if is_sample_csv
-                    else """
-                    If you don't have a CSV on hand to work with,
-                    quit and restart with `dp-wizard --sample`,
-                    and DP Wizard will provide a sample CSV
-                    for the tutorial.
-                    """
-                ),
-                responsive=False,
-            ),
-            ui.row(
-                ui.input_file(
-                    "private_csv_path",
-                    "Choose Private CSV",
-                    accept=[".csv"],
-                    placeholder=Path(initial_private_csv_path).name,
-                ),
-                ui.input_file(
-                    "public_csv_path",
-                    "Choose Public CSV",
-                    accept=[".csv"],
-                    placeholder=Path(initial_public_csv_path).name,
-                ),
-            ),
-        ]
-
-    @render.ui
-    def csv_column_match_ui():
-        mismatch = csv_column_mismatch_calc()
-        messages = []
-        if mismatch:
-            just_public, just_private = mismatch
-            if just_public:
-                messages.append(
-                    "- Only the public CSV contains: "
-                    + ", ".join(f"`{name}`" for name in just_public)
-                )
-            if just_private:
-                messages.append(
-                    "- Only the private CSV contains: "
-                    + ", ".join(f"`{name}`" for name in just_private)
-                )
-        return hide_if(not messages, info_md_box("\n".join(messages)))
 
     entities = {
         "📅 Individual Per Period": """
@@ -453,7 +267,8 @@ Choose both **Private CSV** and **Public CSV** {PUBLIC_PRIVATE_TEXT}
             contributions_valid()
             and not get_row_count_errors(max_rows())
             and len(column_names()) > 0
-            and (in_cloud or not csv_column_mismatch_calc())
+            # TODO: Do per-card validation and re-enable:
+            # and (in_cloud or not csv_column_mismatch_calc())
         )
 
     @reactive.calc
@@ -494,49 +309,11 @@ Choose both **Private CSV** and **Public CSV** {PUBLIC_PRIVATE_TEXT}
             responsive=False,
         )
 
-    @reactive.effect
-    @reactive.event(input.max_rows)
-    def _on_max_rows_change():
-        max_rows.set(input.max_rows())
-
     @render.ui
     def optional_row_count_error_ui():
         error_md = "\n".join(f"- {error}" for error in get_row_count_errors(max_rows()))
         if error_md:
             return info_md_box(error_md)
-
-    @render.ui
-    def row_count_bounds_ui():
-        return (
-            ui.markdown("What is the **maximum row count** of your CSV?"),
-            tutorial_box(
-                is_tutorial_mode(),
-                """
-                If you're unsure, pick a safe value, like the total
-                population of the group being analyzed.
-
-                This value is used downstream two ways:
-                - There is a very small probability that data could be
-                    released verbatim. If your dataset is particularly
-                    large, this probability should be even smaller.
-                - The floating point numbers used by computers are not the
-                    same as the real numbers of mathematics, and with very
-                    large datasets, this gap accumulates, and more noise is
-                    necessary.
-                """,
-                responsive=False,
-            ),
-            ui.layout_columns(
-                ui.input_text(
-                    "max_rows",
-                    only_for_screenreader("Maximum number of rows in CSV"),
-                    "0",
-                ),
-                [],  # column placeholder
-                col_widths=col_widths,  # type: ignore
-            ),
-            ui.output_ui("optional_row_count_error_ui"),
-        )
 
     @render.ui
     def define_analysis_button_ui():
@@ -603,3 +380,5 @@ Choose both **Private CSV** and **Public CSV** {PUBLIC_PRIVATE_TEXT}
     @reactive.event(input.go_to_analysis)
     def go_to_analysis():
         ui.update_navs("top_level_nav", selected="analysis_panel")
+
+    data_source_server(input, output, session, state)
